Исправлены ошибки - отсуствие Исключчения 405, Дубли в ExamServiceImpl и др мелкие помарки

+++ src/main/java/pro/sky/Course2/Services/MathQuestionService.java
@@ -3,6 +3,7 @@ package pro.sky.Course2.Services;
 import org.junit.platform.commons.util.StringUtils;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
+import pro.sky.Course2.Exceptions.AddException;
 import pro.sky.Course2.Exceptions.AlreadyAddException;
 import pro.sky.Course2.Exceptions.EmptyRequestException;
 import pro.sky.Course2.Question;
@@ -21,11 +22,13 @@ public class MathQuestionService implements QuestionService {
     public MathQuestionService(JavaQuestionRepository javaQuestionRepository) {
         this.javaQuestionRepository = javaQuestionRepository;
     }
-int size=4;
-
+    private final int size=4;
+private final int Key=1;
 
     @Override
     public Question add(String query, String answer) {
+        if(Key==1){throw new AddException();
+        }
         if (StringUtils.isBlank(query) == true || StringUtils.isBlank(answer) == true) {
             throw new EmptyRequestException();
         }
@@ -36,6 +39,8 @@ int size=4;
 
     @Override
     public Question add(Question question) {
+        if(Key==1){throw new AddException();
+        }
         if (StringUtils.isBlank(question.getQuery()) == true || StringUtils.isBlank(question.getAnswer()) == true) {
             throw new EmptyRequestException();
         }
@@ -45,12 +50,16 @@ int size=4;
 
     @Override
     public Question remove(Question question) {
+        if(Key==1){throw new AddException();
+        }
         javaQuestionRepository.remove(question);
         return question;
     }
 
     @Override
     public Collection<Question> getAll() {
+        if(Key==1){throw new AddException();
+        }
         return javaQuestionRepository.getAll();
     }
 @Override
